requirements.txt
asgiref==3.3.4
Django==3.2.3
django-filter==2.3.0
django-redis==4.12.1
djangorestframework==3.12.4
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
psycopg2-binary==2.8.5
django-filter==2.3.0
Pillow==8.1.1
